feat: topics for blog posts, static opensource page, dark mode syntax highlighting fixes
