+++ src/app/blog/posts/crossview-master-deployment.mdx
@@ -2,6 +2,7 @@
 title: 'Mastering Crossview Deployment: Securing Your Crossplane Dashboard in an Enterprise Kubernetes Environment'
 publishedAt: '2026-02-05'
 summary: 'Mastering deployment of crossview dashboard covering several authentication mode configuration(none,session,header)'
+topics: 'Kubernetes, DevOps, Security'
 ---
 
 <p align="center">
