+++ src/app/blog/posts/event-driven-architecture.mdx
@@ -2,6 +2,7 @@
 title: 'Event-Driven Architecture for Scalable Platforms'
 publishedAt: '2023-11-14'
 summary: 'A practical overview of event-driven patterns, message contracts, and observability for distributed systems.'
+topics: 'Architecture, Backend, Distributed Systems'
 ---
 
 Event-driven architecture helps teams decouple services while keeping systems responsive under load. Instead of direct synchronous calls everywhere, services publish events that other services consume asynchronously.
