+++ src/app/blog/posts/kubernetes-cost-optimization.mdx
@@ -2,6 +2,7 @@
 title: 'Kubernetes Cost Optimization Without Performance Loss'
 publishedAt: '2022-12-18'
 summary: 'A practical checklist for rightsizing clusters, reducing waste, and improving cost visibility in Kubernetes.'
+topics: 'Kubernetes, DevOps, Cloud'
 ---
 
 Cloud cost optimization in Kubernetes is usually about better defaults, not aggressive downsizing.
