+++ src/app/blog/posts/platform-engineering-roadmap.mdx
@@ -2,6 +2,7 @@
 title: 'Platform Engineering Roadmap for Growing Teams'
 publishedAt: '2023-07-02'
 summary: 'How to introduce platform engineering in stages, from developer pain mapping to reusable internal products.'
+topics: 'Platform Engineering, DevOps'
 ---
 
 Platform engineering is most effective when it solves concrete developer pain first. Start small, measure impact, then expand to broader internal capabilities.
