+++ src/app/blog/posts/spaces-vs-tabs.mdx
@@ -2,6 +2,7 @@
 title: 'Spaces vs. Tabs: The Indentation Debate Continues'
 publishedAt: '2024-04-08'
 summary: 'Explore the enduring debate between using spaces and tabs for code indentation, and why this choice matters more than you might think.'
+topics: 'Opinion, Tooling'
 ---
 
 The debate between using spaces and tabs for indentation in coding may seem trivial to the uninitiated, but it is a topic that continues to inspire passionate discussions among developers. This seemingly minor choice can affect code readability, maintenance, and even team dynamics.
