+++ src/app/blog/posts/static-typing.mdx
@@ -2,6 +2,7 @@
 title: 'The Power of Static Typing in Programming'
 publishedAt: '2024-04-07'
 summary: 'In the ever-evolving landscape of software development, the debate between dynamic and static typing continues to be a hot topic.'
+topics: 'Opinion, Programming'
 ---
 
 In the ever-evolving landscape of software development, the debate between dynamic and static typing continues to be a hot topic. While dynamic typing offers flexibility and rapid development, static typing brings its own set of powerful advantages that can significantly improve the quality and maintainability of code. In this post, we'll explore why static typing is crucial for developers, accompanied by practical examples through markdown code snippets.
