+++ src/app/blog/posts/vim.mdx
@@ -2,6 +2,7 @@
 title: 'Embracing Vim: The Unsung Hero of Code Editors'
 publishedAt: '2024-04-09'
 summary: 'Discover why Vim, with its steep learning curve, remains a beloved tool among developers for editing code efficiently and effectively.'
+topics: 'Tooling, Opinion'
 ---
 
 In the world of software development, where the latest and greatest tools frequently capture the spotlight, Vim stands out as a timeless classic. Despite its age and initial complexity, Vim has managed to retain a devoted following of developers who swear by its efficiency, versatility, and power.
