+++ src/components/ui/provider.tsx
@@ -1,6 +1,6 @@
 "use client"
 
-import { ChakraProvider, defaultSystem } from "@chakra-ui/react"
+import { ChakraProvider, ClientOnly, defaultSystem } from "@chakra-ui/react"
 import {
   ColorModeProvider,
   type ColorModeProviderProps,
@@ -9,7 +9,9 @@ import {
 export function Provider(props: ColorModeProviderProps) {
   return (
     <ChakraProvider value={defaultSystem}>
-      <ColorModeProvider {...props} />
+      <ClientOnly>
+        <ColorModeProvider {...props} />
+      </ClientOnly>
     </ChakraProvider>
   )
 }
